add interfaces for network apis
